refactor out useClickOutside and included it in hooks example page
