+++ apps/next/hooks/useClickOutside.ts
@@ -1,0 +1,19 @@
+import { useEffect, useRef } from 'react'
+
+export function useClickOutside<T extends HTMLElement>(callback: () => void) {
+	const ref = useRef<T>(null)
+
+	useEffect(() => {
+		const handleClickOutside = (event: MouseEvent) => {
+			if (!ref.current?.contains(event.target as Node)) {
+				callback()
+			}
+		}
+		document.addEventListener('click', handleClickOutside)
+		return () => {
+			document.removeEventListener('click', handleClickOutside)
+		}
+	}, [ref, callback])
+
+	return ref
+}
